Promote single values to slice values for weakly typed input
This automatically promotes single standalone values to slices/arrays
if the target type is a slice or array. This only happens in weakly
typed mode.

The single value is weakly decoded into the target slice type, meaning
that a string value can decode to an int slice for example.

This enables the following to decode:

```
source := map[string]interface{}{"key": "4"}

var target struct {
  Key []int
}

// Decode source to target

// Pseudocode
result == { Key: []int{4} }
```