Support squashing maps into structs
This adds support for squashing maps into structs using the `squash`
option.

    type Foo struct {
        Data map[string]interface{} `config:",squash"`
        Bar  string
    }

The squashed map receives fields that were not consumed by the struct.
At most one map may be squashed into a struct. Multiple structs may be
squashed into the struct in addition to the map, in which case the map
will receive fields that were not consumed by all structs combined.