Improve decoding of interface{} fields
A field of type interface{} pointing to an existing value should respect
the value type and fill in the pointed object instead of reassigning the
field with a map.

Thanks to @EaseWay for the original fix:

https://github.com/easeway/mapstructure/commit/3f04f074b18a58796031f043c5529147fac2c707

This commit updates EaseWay's fix and adds an additional test
(TestNestedTypeInterface). This test fails on 3f04f07 but now succeeds
due to additional improvements made since the fix was submitted.

Fixes issue #46.