Check input for nil before and after decode hook.
If input data is nil before decoding, we bail out early, but if decoding
a non-nil value returns a nil value, we continue with the decode method
and return a zero value. As described in #97, this makes it impossible
to return a nil value from a decode hook. This patch checks for nil
inputs before and after decoding.

[Fixes #97]