Add a value hook func
This commit adds a third hook func type which has access to the
reflect.Value representations of both the `from` and the `to`
objects. This gives all the same power of the existing hook functions
(types, kinds, and data can all be retrieved from values), but
with the additional benefit of being able to modify the `to` object.

This change is largely backwards compatible. The only change of the
public API (other than adding a new hook type) is a change in the
signature of `DecodeHookExec`.