Add a recovering decode hook
Hooks are difficult to get 100% correct. When there is an error in the
code using reflection, we are likely to trigger a panic. While a panic
is helpful to the developer, an end-user has no clue where the problem
is and they don't know what part of the configuration would need to be
tweaked to work around the issue.

The proposed recovering decode hook will turn the panic into an error,
which is a more user-friendly way to report the error to an end user.